thêm chức năng thêm ngành
thêm chức năng thêm ngành

+++ database/migrations/2015_11_25_025939_create_nganhs_table.php
@@ -16,7 +16,7 @@ class CreateNganhsTable extends Migration
             $table->increments('id');
             $table->string('manganh');
             $table->string('tennganh');
-            $table->integer('diemchuan');
+            $table->float('diemchuan');
             $table->integer('truong_id')->unsigned();
             $table->foreign('truong_id')->references('id')->on('truongs')->onDelete('cascade');
             $table->timestamps();
